@@ -1,263 +1,121 @@
-//
-// Copyright(c) 2015 Gabi Melman.
+// Copyright(c) 2015-present, Gabi Melman & spdlog contributors.
 // Distributed under the MIT License (http://opensource.org/licenses/MIT)
-//
 
 #pragma once
 
-// Loggers registy of unique name->logger pointer
-// An attempt to create a logger with an already existing name will be ignored
+// Loggers registry of unique name->logger pointer
+// An attempt to create a logger with an already existing name will result with spdlog_ex exception.
 // If user requests a non existing logger, nullptr will be returned
 // This class is thread safe
 
-#include "../async_logger.h"
-#include "../common.h"
-#include "../details/null_mutex.h"
-#include "../logger.h"
+#include <spdlog/common.h>
+#include <spdlog/details/periodic_worker.h>
 
 #include <chrono>
 #include <functional>
 #include <memory>
-#include <mutex>
 #include <string>
 #include <unordered_map>
+#include <mutex>
 
 namespace spdlog {
+class logger;
+
 namespace details {
-template<class Mutex>
-class registry_t
+class thread_pool;
+
+class SPDLOG_API registry
 {
 public:
-    registry_t<Mutex>(const registry_t<Mutex> &) = delete;
-    registry_t<Mutex> &operator=(const registry_t<Mutex> &) = delete;
+    using log_levels = std::unordered_map<std::string, level::level_enum>;
+    registry(const registry &) = delete;
+    registry &operator=(const registry &) = delete;
 
-    void register_logger(std::shared_ptr<logger> logger)
+    void register_logger(std::shared_ptr<logger> new_logger);
+    void initialize_logger(std::shared_ptr<logger> new_logger);
+    std::shared_ptr<logger> get(const std::string &logger_name);
+    std::shared_ptr<logger> default_logger();
+
+    // Return raw ptr to the default logger.
+    // To be used directly by the spdlog default api (e.g. spdlog::info)
+    // This make the default API faster, but cannot be used concurrently with set_default_logger().
+    // e.g do not call set_default_logger() from one thread while calling spdlog::info() from another.
+    logger *get_default_raw();
+
+    // set default logger.
+    // default logger is stored in default_logger_ (for faster retrieval) and in the loggers_ map.
+    void set_default_logger(std::shared_ptr<logger> new_default_logger);
+
+    void set_tp(std::shared_ptr<thread_pool> tp);
+
+    std::shared_ptr<thread_pool> get_tp();
+
+    // Set global formatter. Each sink in each logger will get a clone of this object
+    void set_formatter(std::unique_ptr<formatter> formatter);
+
+    void enable_backtrace(size_t n_messages);
+
+    void disable_backtrace();
+
+    void set_level(level::level_enum log_level);
+
+    void flush_on(level::level_enum log_level);
+
+    template<typename Rep, typename Period>
+    void flush_every(std::chrono::duration<Rep, Period> interval)
     {
-        std::lock_guard<Mutex> lock(_mutex);
-        auto logger_name = logger->name();
-        throw_if_exists(logger_name);
-        _loggers[logger_name] = logger;
+        std::lock_guard<std::mutex> lock(flusher_mutex_);
+        auto clbk = [this]() { this->flush_all(); };
+        periodic_flusher_ = details::make_unique<periodic_worker>(clbk, interval);
     }
 
-    std::shared_ptr<logger> get(const std::string &logger_name)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        auto found = _loggers.find(logger_name);
-        return found == _loggers.end() ? nullptr : found->second;
-    }
+    void set_error_handler(err_handler handler);
 
-    template<class It>
-    std::shared_ptr<logger> create(const std::string &logger_name, const It &sinks_begin, const It &sinks_end)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        throw_if_exists(logger_name);
-        std::shared_ptr<logger> new_logger;
-        if (_async_mode)
-        {
-            new_logger = std::make_shared<async_logger>(logger_name, sinks_begin, sinks_end, _async_q_size, _overflow_policy,
-                _worker_warmup_cb, _flush_interval_ms, _worker_teardown_cb);
-        }
-        else
-        {
-            new_logger = std::make_shared<logger>(logger_name, sinks_begin, sinks_end);
-        }
+    void apply_all(const std::function<void(const std::shared_ptr<logger>)> &fun);
 
-        if (_formatter)
-        {
-            new_logger->set_formatter(_formatter);
-        }
+    void flush_all();
 
-        if (_err_handler)
-        {
-            new_logger->set_error_handler(_err_handler);
-        }
+    void drop(const std::string &logger_name);
 
-        new_logger->set_level(_level);
-        new_logger->flush_on(_flush_level);
+    void drop_all();
 
-        // Add to registry
-        _loggers[logger_name] = new_logger;
-        return new_logger;
-    }
+    // clean all resources and threads started by the registry
+    void shutdown();
 
-    template<class It>
-    std::shared_ptr<async_logger> create_async(const std::string &logger_name, size_t queue_size,
-        const async_overflow_policy overflow_policy, const std::function<void()> &worker_warmup_cb,
-        const std::chrono::milliseconds &flush_interval_ms, const std::function<void()> &worker_teardown_cb, const It &sinks_begin,
-        const It &sinks_end)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        throw_if_exists(logger_name);
-        auto new_logger = std::make_shared<async_logger>(
-            logger_name, sinks_begin, sinks_end, queue_size, overflow_policy, worker_warmup_cb, flush_interval_ms, worker_teardown_cb);
+    std::recursive_mutex &tp_mutex();
 
-        if (_formatter)
-        {
-            new_logger->set_formatter(_formatter);
-        }
+    void set_automatic_registration(bool automatic_registration);
 
-        if (_err_handler)
-        {
-            new_logger->set_error_handler(_err_handler);
-        }
+    // set levels for all existing/future loggers. global_level can be null if should not set.
+    void set_levels(log_levels levels, level::level_enum *global_level);
 
-        new_logger->set_level(_level);
-        new_logger->flush_on(_flush_level);
-
-        // Add to registry
-        _loggers[logger_name] = new_logger;
-        return new_logger;
-    }
-
-    void apply_all(std::function<void(std::shared_ptr<logger>)> fun)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        for (auto &l : _loggers)
-        {
-            fun(l.second);
-        }
-    }
-
-    void drop(const std::string &logger_name)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _loggers.erase(logger_name);
-    }
-
-    void drop_all()
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _loggers.clear();
-    }
-
-    std::shared_ptr<logger> create(const std::string &logger_name, sinks_init_list sinks)
-    {
-        return create(logger_name, sinks.begin(), sinks.end());
-    }
-
-    std::shared_ptr<logger> create(const std::string &logger_name, sink_ptr sink)
-    {
-        return create(logger_name, {sink});
-    }
-
-    std::shared_ptr<async_logger> create_async(const std::string &logger_name, size_t queue_size,
-        const async_overflow_policy overflow_policy, const std::function<void()> &worker_warmup_cb,
-        const std::chrono::milliseconds &flush_interval_ms, const std::function<void()> &worker_teardown_cb, sinks_init_list sinks)
-    {
-        return create_async(
-            logger_name, queue_size, overflow_policy, worker_warmup_cb, flush_interval_ms, worker_teardown_cb, sinks.begin(), sinks.end());
-    }
-
-    std::shared_ptr<async_logger> create_async(const std::string &logger_name, size_t queue_size,
-        const async_overflow_policy overflow_policy, const std::function<void()> &worker_warmup_cb,
-        const std::chrono::milliseconds &flush_interval_ms, const std::function<void()> &worker_teardown_cb, sink_ptr sink)
-    {
-        return create_async(logger_name, queue_size, overflow_policy, worker_warmup_cb, flush_interval_ms, worker_teardown_cb, {sink});
-    }
-
-    void formatter(formatter_ptr f)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _formatter = f;
-        for (auto &l : _loggers)
-        {
-            l.second->set_formatter(_formatter);
-        }
-    }
-
-    void set_pattern(const std::string &pattern)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _formatter = std::make_shared<pattern_formatter>(pattern);
-        for (auto &l : _loggers)
-        {
-            l.second->set_formatter(_formatter);
-        }
-    }
-
-    void set_level(level::level_enum log_level)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        for (auto &l : _loggers)
-        {
-            l.second->set_level(log_level);
-        }
-        _level = log_level;
-    }
-
-    void flush_on(level::level_enum log_level)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        for (auto &l : _loggers)
-        {
-            l.second->flush_on(log_level);
-        }
-        _flush_level = log_level;
-    }
-
-    void set_error_handler(log_err_handler handler)
-    {
-        for (auto &l : _loggers)
-        {
-            l.second->set_error_handler(handler);
-        }
-        _err_handler = handler;
-    }
-
-    void set_async_mode(size_t q_size, const async_overflow_policy overflow_policy, const std::function<void()> &worker_warmup_cb,
-        const std::chrono::milliseconds &flush_interval_ms, const std::function<void()> &worker_teardown_cb)
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _async_mode = true;
-        _async_q_size = q_size;
-        _overflow_policy = overflow_policy;
-        _worker_warmup_cb = worker_warmup_cb;
-        _flush_interval_ms = flush_interval_ms;
-        _worker_teardown_cb = worker_teardown_cb;
-    }
-
-    void set_sync_mode()
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _async_mode = false;
-    }
-
-    static registry_t<Mutex> &instance()
-    {
-        static registry_t<Mutex> s_instance;
-        return s_instance;
-    }
+    static registry &instance();
 
 private:
-    registry_t<Mutex>() = default;
+    registry();
+    ~registry();
 
-    void throw_if_exists(const std::string &logger_name)
-    {
-        if (_loggers.find(logger_name) != _loggers.end())
-        {
-            throw spdlog_ex("logger with name '" + logger_name + "' already exists");
-        }
-    }
-
-    Mutex _mutex;
-    std::unordered_map<std::string, std::shared_ptr<logger>> _loggers;
-    formatter_ptr _formatter;
-    level::level_enum _level = level::info;
-    level::level_enum _flush_level = level::off;
-    log_err_handler _err_handler;
-    bool _async_mode = false;
-    size_t _async_q_size = 0;
-    async_overflow_policy _overflow_policy = async_overflow_policy::block_retry;
-    std::function<void()> _worker_warmup_cb;
-    std::chrono::milliseconds _flush_interval_ms{std::chrono::milliseconds::zero()};
-    std::function<void()> _worker_teardown_cb;
+    void throw_if_exists_(const std::string &logger_name);
+    void register_logger_(std::shared_ptr<logger> new_logger);
+    bool set_level_from_cfg_(logger *logger);
+    std::mutex logger_map_mutex_, flusher_mutex_;
+    std::recursive_mutex tp_mutex_;
+    std::unordered_map<std::string, std::shared_ptr<logger>> loggers_;
+    log_levels log_levels_;
+    std::unique_ptr<formatter> formatter_;
+    spdlog::level::level_enum global_log_level_ = level::info;
+    level::level_enum flush_level_ = level::off;
+    err_handler err_handler_;
+    std::shared_ptr<thread_pool> tp_;
+    std::unique_ptr<periodic_worker> periodic_flusher_;
+    std::shared_ptr<logger> default_logger_;
+    bool automatic_registration_ = true;
+    size_t backtrace_n_messages_ = 0;
 };
-
-#ifdef SPDLOG_NO_REGISTRY_MUTEX
-using registry = registry_t<spdlog::details::null_mutex>;
-#else
-using registry = registry_t<std::mutex>;
-#endif
 
 } // namespace details
 } // namespace spdlog
+
+#ifdef SPDLOG_HEADER_ONLY
+#    include "registry-inl.h"
+#endif
