@@ -1,150 +1,118 @@
-//
-// Copyright(c) 2017 spdlog authors.
+// Copyright(c) 2015-present, Gabi Melman & spdlog contributors.
 // Distributed under the MIT License (http://opensource.org/licenses/MIT)
-//
 
 #pragma once
 
-#include "../common.h"
-#include "../details/os.h"
-#include "base_sink.h"
-
+#include <spdlog/details/console_globals.h>
+#include <spdlog/details/null_mutex.h>
+#include <spdlog/sinks/sink.h>
+#include <memory>
+#include <mutex>
 #include <string>
-#include <unordered_map>
+#include <array>
 
 namespace spdlog {
 namespace sinks {
 
 /**
- * This sink prefixes the output with an ANSI escape sequence color code depending on the severity
+ * This sink prefixes the output with an ANSI escape sequence color code
+ * depending on the severity
  * of the message.
  * If no color terminal detected, omit the escape codes.
  */
-template<class Mutex>
-class ansicolor_sink : public base_sink<Mutex>
+
+template<typename ConsoleMutex>
+class ansicolor_sink : public sink
 {
 public:
-    explicit ansicolor_sink(FILE *file)
-        : target_file_(file)
-    {
-        should_do_colors_ = details::os::in_terminal(file) && details::os::is_color_terminal();
-        colors_[level::trace] = white;
-        colors_[level::debug] = cyan;
-        colors_[level::info] = green;
-        colors_[level::warn] = yellow + bold;
-        colors_[level::err] = red + bold;
-        colors_[level::critical] = bold + on_red;
-        colors_[level::off] = reset;
-    }
+    using mutex_t = typename ConsoleMutex::mutex_t;
+    ansicolor_sink(FILE *target_file, color_mode mode);
+    ~ansicolor_sink() override = default;
 
-    ~ansicolor_sink() override
-    {
-        _flush();
-    }
+    ansicolor_sink(const ansicolor_sink &other) = delete;
+    ansicolor_sink(ansicolor_sink &&other) = delete;
 
-    void set_color(level::level_enum color_level, const std::string &color)
-    {
-        std::lock_guard<Mutex> lock(base_sink<Mutex>::_mutex);
-        colors_[color_level] = color;
-    }
+    ansicolor_sink &operator=(const ansicolor_sink &other) = delete;
+    ansicolor_sink &operator=(ansicolor_sink &&other) = delete;
 
-    /// Formatting codes
-    const std::string reset = "\033[m";
-    const std::string bold = "\033[1m";
-    const std::string dark = "\033[2m";
-    const std::string underline = "\033[4m";
-    const std::string blink = "\033[5m";
-    const std::string reverse = "\033[7m";
-    const std::string concealed = "\033[8m";
-    const std::string clear_line = "\033[K";
+    void set_color(level::level_enum color_level, string_view_t color);
+    void set_color_mode(color_mode mode);
+    bool should_color();
+
+    void log(const details::log_msg &msg) override;
+    void flush() override;
+    void set_pattern(const std::string &pattern) final;
+    void set_formatter(std::unique_ptr<spdlog::formatter> sink_formatter) override;
+
+    // Formatting codes
+    const string_view_t reset = "\033[m";
+    const string_view_t bold = "\033[1m";
+    const string_view_t dark = "\033[2m";
+    const string_view_t underline = "\033[4m";
+    const string_view_t blink = "\033[5m";
+    const string_view_t reverse = "\033[7m";
+    const string_view_t concealed = "\033[8m";
+    const string_view_t clear_line = "\033[K";
 
     // Foreground colors
-    const std::string black = "\033[30m";
-    const std::string red = "\033[31m";
-    const std::string green = "\033[32m";
-    const std::string yellow = "\033[33m";
-    const std::string blue = "\033[34m";
-    const std::string magenta = "\033[35m";
-    const std::string cyan = "\033[36m";
-    const std::string white = "\033[37m";
+    const string_view_t black = "\033[30m";
+    const string_view_t red = "\033[31m";
+    const string_view_t green = "\033[32m";
+    const string_view_t yellow = "\033[33m";
+    const string_view_t blue = "\033[34m";
+    const string_view_t magenta = "\033[35m";
+    const string_view_t cyan = "\033[36m";
+    const string_view_t white = "\033[37m";
 
     /// Background colors
-    const std::string on_black = "\033[40m";
-    const std::string on_red = "\033[41m";
-    const std::string on_green = "\033[42m";
-    const std::string on_yellow = "\033[43m";
-    const std::string on_blue = "\033[44m";
-    const std::string on_magenta = "\033[45m";
-    const std::string on_cyan = "\033[46m";
-    const std::string on_white = "\033[47m";
+    const string_view_t on_black = "\033[40m";
+    const string_view_t on_red = "\033[41m";
+    const string_view_t on_green = "\033[42m";
+    const string_view_t on_yellow = "\033[43m";
+    const string_view_t on_blue = "\033[44m";
+    const string_view_t on_magenta = "\033[45m";
+    const string_view_t on_cyan = "\033[46m";
+    const string_view_t on_white = "\033[47m";
 
-protected:
-    void _sink_it(const details::log_msg &msg) override
-    {
-        // Wrap the originally formatted message in color codes.
-        // If color is not supported in the terminal, log as is instead.
-        if (should_do_colors_ && msg.color_range_end > msg.color_range_start)
-        {
-            // before color range
-            _print_range(msg, 0, msg.color_range_start);
-            // in color range
-            _print_ccode(colors_[msg.level]);
-            _print_range(msg, msg.color_range_start, msg.color_range_end);
-            _print_ccode(reset);
-            // after color range
-            _print_range(msg, msg.color_range_end, msg.formatted.size());
-        }
-        else
-        {
-            _print_range(msg, 0, msg.formatted.size());
-        }
-        _flush();
-    }
-
-    void _flush() override
-    {
-        fflush(target_file_);
-    }
+    /// Bold colors
+    const string_view_t yellow_bold = "\033[33m\033[1m";
+    const string_view_t red_bold = "\033[31m\033[1m";
+    const string_view_t bold_on_red = "\033[1m\033[41m";
 
 private:
-    void _print_ccode(const std::string &color_code)
-    {
-        fwrite(color_code.data(), sizeof(char), color_code.size(), target_file_);
-    }
-    void _print_range(const details::log_msg &msg, size_t start, size_t end)
-    {
-        fwrite(msg.formatted.data() + start, sizeof(char), end - start, target_file_);
-    }
     FILE *target_file_;
+    mutex_t &mutex_;
     bool should_do_colors_;
-    std::unordered_map<level::level_enum, std::string, level::level_hasher> colors_;
+    std::unique_ptr<spdlog::formatter> formatter_;
+    std::array<std::string, level::n_levels> colors_;
+    void print_ccode_(const string_view_t &color_code);
+    void print_range_(const memory_buf_t &formatted, size_t start, size_t end);
+    static std::string to_string_(const string_view_t &sv);
 };
 
-template<class Mutex>
-class ansicolor_stdout_sink : public ansicolor_sink<Mutex>
+template<typename ConsoleMutex>
+class ansicolor_stdout_sink : public ansicolor_sink<ConsoleMutex>
 {
 public:
-    ansicolor_stdout_sink()
-        : ansicolor_sink<Mutex>(stdout)
-    {
-    }
+    explicit ansicolor_stdout_sink(color_mode mode = color_mode::automatic);
 };
 
-using ansicolor_stdout_sink_mt = ansicolor_stdout_sink<std::mutex>;
-using ansicolor_stdout_sink_st = ansicolor_stdout_sink<details::null_mutex>;
-
-template<class Mutex>
-class ansicolor_stderr_sink : public ansicolor_sink<Mutex>
+template<typename ConsoleMutex>
+class ansicolor_stderr_sink : public ansicolor_sink<ConsoleMutex>
 {
 public:
-    ansicolor_stderr_sink()
-        : ansicolor_sink<Mutex>(stderr)
-    {
-    }
+    explicit ansicolor_stderr_sink(color_mode mode = color_mode::automatic);
 };
 
-using ansicolor_stderr_sink_mt = ansicolor_stderr_sink<std::mutex>;
-using ansicolor_stderr_sink_st = ansicolor_stderr_sink<details::null_mutex>;
+using ansicolor_stdout_sink_mt = ansicolor_stdout_sink<details::console_mutex>;
+using ansicolor_stdout_sink_st = ansicolor_stdout_sink<details::console_nullmutex>;
+
+using ansicolor_stderr_sink_mt = ansicolor_stderr_sink<details::console_mutex>;
+using ansicolor_stderr_sink_st = ansicolor_stderr_sink<details::console_nullmutex>;
 
 } // namespace sinks
 } // namespace spdlog
+
+#ifdef SPDLOG_HEADER_ONLY
+#    include "ansicolor_sink-inl.h"
+#endif
