@@ -1,49 +1,52 @@
-//
-// Copyright(c) 2015 Gabi Melman.
+// Copyright(c) 2015-present, Gabi Melman & spdlog contributors.
 // Distributed under the MIT License (http://opensource.org/licenses/MIT)
-//
 
 #pragma once
 //
 // base sink templated over a mutex (either dummy or real)
-// concrete implementation should only override the _sink_it method.
-// all locking is taken care of here so no locking needed by the implementers..
+// concrete implementation should override the sink_it_() and flush_()  methods.
+// locking is taken care of in this class - no locking needed by the
+// implementers..
 //
 
-#include "../common.h"
-#include "../details/log_msg.h"
-#include "../formatter.h"
-#include "sink.h"
-
-#include <mutex>
+#include <spdlog/common.h>
+#include <spdlog/details/log_msg.h>
+#include <spdlog/sinks/sink.h>
 
 namespace spdlog {
 namespace sinks {
-template<class Mutex>
-class base_sink : public sink
+template<typename Mutex>
+class SPDLOG_API base_sink : public sink
 {
 public:
-    base_sink() = default;
+    base_sink();
+    explicit base_sink(std::unique_ptr<spdlog::formatter> formatter);
+    ~base_sink() override = default;
 
     base_sink(const base_sink &) = delete;
+    base_sink(base_sink &&) = delete;
+
     base_sink &operator=(const base_sink &) = delete;
+    base_sink &operator=(base_sink &&) = delete;
 
-    void log(const details::log_msg &msg) SPDLOG_FINAL override
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _sink_it(msg);
-    }
-
-    void flush() SPDLOG_FINAL override
-    {
-        std::lock_guard<Mutex> lock(_mutex);
-        _flush();
-    }
+    void log(const details::log_msg &msg) final;
+    void flush() final;
+    void set_pattern(const std::string &pattern) final;
+    void set_formatter(std::unique_ptr<spdlog::formatter> sink_formatter) final;
 
 protected:
-    virtual void _sink_it(const details::log_msg &msg) = 0;
-    virtual void _flush() = 0;
-    Mutex _mutex;
+    // sink formatter
+    std::unique_ptr<spdlog::formatter> formatter_;
+    Mutex mutex_;
+
+    virtual void sink_it_(const details::log_msg &msg) = 0;
+    virtual void flush_() = 0;
+    virtual void set_pattern_(const std::string &pattern);
+    virtual void set_formatter_(std::unique_ptr<spdlog::formatter> sink_formatter);
 };
 } // namespace sinks
 } // namespace spdlog
+
+#ifdef SPDLOG_HEADER_ONLY
+#    include "base_sink-inl.h"
+#endif
