@@ -1,42 +1,47 @@
-//
 // Copyright(c) 2016 Alexander Dalshov.
 // Distributed under the MIT License (http://opensource.org/licenses/MIT)
-//
 
 #pragma once
 
 #if defined(_WIN32)
 
-#include "../details/null_mutex.h"
-#include "base_sink.h"
+#    include <spdlog/details/null_mutex.h>
+#    include <spdlog/sinks/base_sink.h>
 
-#include <winbase.h>
+#    include <mutex>
+#    include <string>
 
-#include <mutex>
-#include <string>
+// Avoid including windows.h (https://stackoverflow.com/a/30741042)
+extern "C" __declspec(dllimport) void __stdcall OutputDebugStringA(const char *lpOutputString);
 
 namespace spdlog {
 namespace sinks {
 /*
  * MSVC sink (logging using OutputDebugStringA)
  */
-template<class Mutex>
+template<typename Mutex>
 class msvc_sink : public base_sink<Mutex>
 {
 public:
-    explicit msvc_sink() {}
+    msvc_sink() = default;
 
 protected:
-    void _sink_it(const details::log_msg &msg) override
+    void sink_it_(const details::log_msg &msg) override
     {
-        OutputDebugStringA(msg.formatted.c_str());
+        memory_buf_t formatted;
+        base_sink<Mutex>::formatter_->format(msg, formatted);
+        formatted.push_back('\0'); // add a null terminator for OutputDebugStringA
+        OutputDebugStringA(formatted.data());
     }
 
-    void _flush() override {}
+    void flush_() override {}
 };
 
 using msvc_sink_mt = msvc_sink<std::mutex>;
 using msvc_sink_st = msvc_sink<details::null_mutex>;
+
+using windebug_sink_mt = msvc_sink_mt;
+using windebug_sink_st = msvc_sink_st;
 
 } // namespace sinks
 } // namespace spdlog
