@@ -6,12 +6,22 @@
 #include <cstdio>
 #include <exception>
 #include <fstream>
+#include <iostream>
 #include <ostream>
+#include <sstream>
 #include <string>
+#include <iomanip>
+#include <stdlib.h>
 
-#define SPDLOG_TRACE_ON
-#define SPDLOG_DEBUG_ON
+#define SPDLOG_ACTIVE_LEVEL SPDLOG_LEVEL_DEBUG
 
-#include "../include/spdlog/sinks/null_sink.h"
-#include "../include/spdlog/sinks/ostream_sink.h"
-#include "../include/spdlog/spdlog.h"
+#include "spdlog/spdlog.h"
+#include "spdlog/async.h"
+#include "spdlog/details/fmt_helper.h"
+#include "spdlog/sinks/basic_file_sink.h"
+#include "spdlog/sinks/daily_file_sink.h"
+#include "spdlog/sinks/null_sink.h"
+#include "spdlog/sinks/ostream_sink.h"
+#include "spdlog/sinks/rotating_file_sink.h"
+#include "spdlog/sinks/stdout_color_sinks.h"
+#include "spdlog/pattern_formatter.h"
